Completed HW1 and added Common
Common contains some library functions while HW1 is the Kaleidescope
assignment. It required us to create a piece of geometry centered at
the origin and then draw 10 of the same geometry in a circle around
it. Then we were needed to rotate the origin geometry clockwise and
the circle counter-clockwise. Finally, we had to create some sort of
function that altered the color of the geometry based on the coordinates.